removed api.ts, renamed file to admin.ts and fixed imports for Competition.tsx, Users.tsx, admin,ts
